feat: review nns and sns neuron details (#1553)
# Motivation

Review nns and sns neuron details information structure - e.g. display order and action grouping.

# Note

- This PR does not contain any functional changes. Code is refactored and display structure is changed but no functional changes.

- This PR contains commented code for the display of the neuron "Age". This will be uncommented and proposed in a separate PR.

# Changes

- refactor few components packages from "neurons/-detail" to "sns-neurons/-detail" and name as well to prefix with "Nns" or "Sns"
- new `<Separator />` component that renders a `hr`
- structure and presentation of the information reviewed
- few component content extracted to dedicated content because the detail page does not use the `NeuronCard` for rendering purpose anymore

# Unrelated
- "Your principal id is" text modified to "Your principal is" to shorten the sentence in accordance with Mischa
