+++ frontend/src/lib/components/neuron-detail/NeuronFollowingCard/NeuronFollowingCard.svelte
@@ -15,6 +15,7 @@
   import FollowNeuronsButton from "../actions/FollowNeuronsButton.svelte";
   import Followee from "./Followee.svelte";
   import { KeyValuePairInfo } from "@dfinity/gix-components";
+  import Separator from "$lib/components/ui/Separator.svelte";
 
   export let neuron: NeuronInfo;
   let isControllable: boolean;
@@ -58,6 +59,8 @@
     {/if}
   </div>
 </CardInfo>
+
+<Separator />
 
 <style lang="scss">
   h3 {
