+++ frontend/src/lib/components/neuron-detail/NeuronMaturityCard.svelte
@@ -19,6 +19,7 @@
     formattedTotalMaturity,
   } from "$lib/utils/neuron.utils";
   import { accountsStore } from "$lib/stores/accounts.store";
+  import Separator from "$lib/components/ui/Separator.svelte";
 
   export let neuron: NeuronInfo;
 
@@ -78,6 +79,8 @@
     {/if}
   {/if}
 </CardInfo>
+
+<Separator />
 
 <style lang="scss">
   h3 {
