+++ frontend/src/lib/components/neurons/ConfirmNeuronsMerge.svelte
@@ -67,7 +67,7 @@
       <p class="label">{$i18n.neurons.neuron_balance}</p>
       <p>
         <Html
-          text={replacePlaceholders($i18n.neurons.icp_stake, {
+          text={replacePlaceholders($i18n.neurons.amount_icp_stake, {
             $amount: valueSpan(
               formatToken({ value: neuronStake(neuron), detailed: true })
             ),
