+++ frontend/src/lib/components/neurons/NeuronStateInfo.svelte
@@ -13,7 +13,7 @@
 </script>
 
 {#if stateInfo !== undefined}
-  <div class="status">
+  <div class="status" data-tid="neuron-state-info">
     <svelte:component this={stateInfo.Icon} />
     {keyOf({ obj: $i18n.neuron_state, key: NeuronState[state] })}
   </div>
@@ -28,7 +28,6 @@
     align-items: center;
 
     border-radius: var(--border-radius-0_5x);
-    padding: var(--padding) var(--padding-2x) var(--padding) var(--padding);
 
     @include fonts.small;
