+++ frontend/src/lib/components/neurons/NeuronStateRemainingTime.svelte
@@ -3,25 +3,47 @@
   import { i18n } from "$lib/stores/i18n";
   import { secondsToDuration } from "$lib/utils/date.utils";
   import { replacePlaceholders } from "$lib/utils/i18n.utils";
-  import { Html } from "@dfinity/gix-components";
+  import { Html, KeyValuePair } from "@dfinity/gix-components";
 
   export let state: NeuronState;
   export let timeInSeconds: bigint | undefined;
+  export let inline = true;
 </script>
 
 {#if timeInSeconds !== undefined}
   {#if state === NeuronState.Dissolving || state === NeuronState.Spawning}
-    <p class="duration description">
-      <Html
-        text={replacePlaceholders($i18n.neurons.remaining, {
-          $duration: secondsToDuration(timeInSeconds),
-        })}
-      />
-    </p>
+    {#if inline}
+      <p class="duration description">
+        <Html
+          text={replacePlaceholders($i18n.neurons.inline_remaining, {
+            $duration: secondsToDuration(timeInSeconds),
+          })}
+        />
+      </p>
+    {:else}
+      <KeyValuePair>
+        <span slot="key" class="label">{$i18n.neurons.remaining}</span>
+        <span slot="value" class="value"
+          >{secondsToDuration(timeInSeconds)}</span
+        >
+      </KeyValuePair>
+    {/if}
   {:else if state === NeuronState.Locked}
-    <p class="duration description">
-      {secondsToDuration(timeInSeconds)} – {$i18n.neurons.dissolve_delay_title}
-    </p>
+    {#if inline}
+      <p class="duration description">
+        {secondsToDuration(timeInSeconds)} – {$i18n.neurons
+          .dissolve_delay_title}
+      </p>
+    {:else}
+      <KeyValuePair>
+        <span slot="key" class="label"
+          >{$i18n.neurons.dissolve_delay_title}</span
+        >
+        <span slot="value" class="value"
+          >{secondsToDuration(timeInSeconds)}</span
+        >
+      </KeyValuePair>
+    {/if}
   {/if}
 {/if}
