+++ frontend/src/lib/components/neurons/SetDissolveDelay.svelte
@@ -56,7 +56,7 @@
     <p class="label">{$i18n.neurons.neuron_balance}</p>
     <p data-tid="neuron-stake">
       <Html
-        text={replacePlaceholders($i18n.neurons.icp_stake, {
+        text={replacePlaceholders($i18n.neurons.amount_icp_stake, {
           $amount: valueSpan(formatToken({ value: neuronICP, detailed: true })),
         })}
       />
