+++ frontend/src/lib/components/proposal-detail/VotingCard/VotingNeuronSelect.svelte
@@ -104,7 +104,7 @@
   }
 
   .icon {
-    color: var(--line);
+    color: var(--tertiary);
     display: flex;
     align-items: center;
     justify-content: center;
