+++ frontend/src/lib/components/sns-neurons/SnsNeuronsFooter.svelte
