+++ frontend/src/lib/components/ui/Separator.svelte
@@ -1,0 +1,8 @@
+<hr />
+
+<style lang="scss">
+  hr {
+    color: var(--line);
+    margin: var(--padding-4x) 0;
+  }
+</style>
