+++ frontend/src/lib/components/ui/SkeletonCard.svelte
@@ -4,9 +4,11 @@
   import { SkeletonText } from "@dfinity/gix-components";
   import type { SvelteComponent } from "svelte";
   import type { CardType } from "$lib/types/card";
+  import Separator from "$lib/components/ui/Separator.svelte";
 
   export let size: "medium" | "large" = "medium";
   export let cardType: CardType = "card";
+  export let separator = false;
 
   const cards: Record<CardType, typeof SvelteComponent> = {
     card: Card,
@@ -32,6 +34,10 @@
     {/if}
   </div>
 </svelte:component>
+
+{#if separator}
+  <Separator />
+{/if}
 
 <style lang="scss">
   .small {
