+++ frontend/src/lib/i18n/en.json
@@ -27,7 +27,7 @@
     "ic": "Internet Computer",
     "previous": "Previous",
     "next": "Next",
-    "principal_is": "Your principal id is"
+    "principal_is": "Your principal is"
   },
   "error": {
     "auth_sync": "There was an unexpected issue while syncing the status of your authentication. Try to refresh your browser.",
@@ -228,9 +228,12 @@
     "community_fund": "Community fund",
     "hotkey_control": "Hotkey control",
     "stake": "Stake",
-    "icp_stake": "$amount ICP Stake",
+    "amount_icp_stake": "$amount ICP Stake",
+    "ic_stake": "ICP Stake",
     "staked": "Staked",
-    "remaining": "$duration remaining",
+    "inline_remaining": "$duration remaining",
+    "remaining": "Remaining",
+    "age": "Age",
     "aria_label_neuron_card": "Go to neuron details",
     "neuron_id": "Neuron ID",
     "neuron_balance": "Balance",
@@ -259,7 +262,7 @@
     "add_hotkey_prompt_success": "Hotkey added successfully.",
     "top_up_neuron": "Increase Neuron Stake",
     "top_up_description": "Increase stake of neuron $neuronId",
-    "community_fund_title": "Community Fund",
+    "community_fund_title": "Community Fund Membership",
     "stake_amount": "Stake Amount"
   },
   "new_followee": {
@@ -459,7 +462,6 @@
     "voting_history": "Voting History",
     "vote": "Vote",
     "join_community_fund": "Join Community Fund",
-    "community_fund": "Community Fund",
     "increase_dissolve_delay": "Increase Delay",
     "start_dissolving": "Start Dissolving",
     "stop_dissolving": "Stop Dissolving",
