+++ frontend/src/lib/types/i18n.d.ts
@@ -238,9 +238,12 @@ interface I18nNeurons {
   community_fund: string;
   hotkey_control: string;
   stake: string;
-  icp_stake: string;
+  amount_icp_stake: string;
+  ic_stake: string;
   staked: string;
+  inline_remaining: string;
   remaining: string;
+  age: string;
   aria_label_neuron_card: string;
   neuron_id: string;
   neuron_balance: string;
@@ -481,7 +484,6 @@ interface I18nNeuron_detail {
   voting_history: string;
   vote: string;
   join_community_fund: string;
-  community_fund: string;
   increase_dissolve_delay: string;
   start_dissolving: string;
   stop_dissolving: string;
