+++ frontend/src/tests/lib/components/neuron-detail/NnsNeuronInfoStake.spec.ts
@@ -1,0 +1,53 @@
+/**
+ * @jest-environment jsdom
+ */
+
+import NnsNeuronInfoStake from "$lib/components/neuron-detail/NnsNeuronInfoStake.svelte";
+import { accountsStore } from "$lib/stores/accounts.store";
+import { authStore } from "$lib/stores/auth.store";
+import { render } from "@testing-library/svelte";
+import {
+  mockAccountsStoreSubscribe,
+  mockHardwareWalletAccount,
+  mockMainAccount,
+} from "../../../mocks/accounts.store.mock";
+import { mockAuthStoreSubscribe } from "../../../mocks/auth.store.mock";
+import en from "../../../mocks/i18n.mock";
+import { mockFullNeuron, mockNeuron } from "../../../mocks/neurons.mock";
+
+describe("NnsNeuronInfoStake", () => {
+  const props = {
+    neuron: {
+      ...mockNeuron,
+      fullNeuron: {
+        ...mockFullNeuron,
+        controller: mockMainAccount.principal?.toText() as string,
+      },
+    },
+  };
+
+  beforeAll(() => {
+    jest
+      .spyOn(authStore, "subscribe")
+      .mockImplementation(mockAuthStoreSubscribe);
+
+    jest
+      .spyOn(accountsStore, "subscribe")
+      .mockImplementation(
+        mockAccountsStoreSubscribe([], [mockHardwareWalletAccount])
+      );
+  });
+
+  it("renders actions", () => {
+    // Each action button is tested separately
+    const { queryByText } = render(NnsNeuronInfoStake, {
+      props,
+    });
+
+    expect(
+      queryByText(en.neuron_detail.increase_dissolve_delay)
+    ).toBeInTheDocument();
+    expect(queryByText(en.neuron_detail.start_dissolving)).toBeInTheDocument();
+    expect(queryByText(en.neuron_detail.increase_stake)).toBeInTheDocument();
+  });
+});
