+++ frontend/src/tests/lib/components/sns-neuron-detail/SnsNeuronMetaInfoCard.spec.ts
@@ -5,13 +5,12 @@
 import SnsNeuronMetaInfoCard from "$lib/components/sns-neuron-detail/SnsNeuronMetaInfoCard.svelte";
 import { snsTokenSymbolSelectedStore } from "$lib/derived/sns/sns-token-symbol-selected.store";
 import { authStore } from "$lib/stores/auth.store";
-import { SnsNeuronPermissionType } from "@dfinity/sns";
+import { shortenWithMiddleEllipsis } from "$lib/utils/format.utils";
+import { getSnsNeuronIdAsHexString } from "$lib/utils/sns-neuron.utils";
 import { mockAuthStoreSubscribe } from "../../../mocks/auth.store.mock";
 import { renderSelectedSnsNeuronContext } from "../../../mocks/context-wrapper.mock";
-import {
-  mockSnsNeuron,
-  mockSnsNeuronWithPermissions,
-} from "../../../mocks/sns-neurons.mock";
+import en from "../../../mocks/i18n.mock";
+import { mockSnsNeuron } from "../../../mocks/sns-neurons.mock";
 import { mockTokenStore } from "../../../mocks/sns-projects.mock";
 
 describe("SnsNeuronMetaInfoCard", () => {
@@ -25,90 +24,41 @@ describe("SnsNeuronMetaInfoCard", () => {
       .mockImplementation(mockAuthStoreSubscribe);
   });
 
-  it("renders a SnsNeuronCard", () => {
-    // We can skip many edge cases tested in the NeuronCard
-    const { queryByTestId } = renderSelectedSnsNeuronContext({
+  it("should render neuron id", () => {
+    const { getByTestId } = renderSelectedSnsNeuronContext({
       Component: SnsNeuronMetaInfoCard,
       neuron: mockSnsNeuron,
       reload: jest.fn(),
     });
 
-    expect(queryByTestId("sns-neuron-card-title")).toBeInTheDocument();
+    const hash = shortenWithMiddleEllipsis(
+      `${getSnsNeuronIdAsHexString(mockSnsNeuron) ?? ""}`
+    );
+    expect(getByTestId("neuron-id")?.textContent.trim()).toEqual(hash);
   });
 
-  it("should render disburse button", async () => {
-    const neuron = mockSnsNeuronWithPermissions([
-      SnsNeuronPermissionType.NEURON_PERMISSION_TYPE_DISBURSE,
-    ]);
-    const { queryByTestId } = renderSelectedSnsNeuronContext({
+  it("should render neuron state", () => {
+    const { getByTestId } = renderSelectedSnsNeuronContext({
       Component: SnsNeuronMetaInfoCard,
-      neuron,
+      neuron: mockSnsNeuron,
       reload: jest.fn(),
     });
 
-    expect(queryByTestId("disburse-button")).toBeInTheDocument();
+    expect(getByTestId("neuron-state-info")?.textContent.trim()).toEqual(
+      en.neuron_state.Dissolved
+    );
   });
 
-  it("should not render disburse button", async () => {
-    const neuron = mockSnsNeuronWithPermissions([]);
-    const { queryByTestId } = renderSelectedSnsNeuronContext({
-      Component: SnsNeuronMetaInfoCard,
-      neuron,
-      reload: jest.fn(),
-    });
-
-    expect(queryByTestId("disburse-button")).not.toBeInTheDocument();
-  });
-
-  it("should render dissolve button", async () => {
-    const neuron = mockSnsNeuronWithPermissions([
-      SnsNeuronPermissionType.NEURON_PERMISSION_TYPE_CONFIGURE_DISSOLVE_STATE,
-    ]);
-    const { queryByTestId } = renderSelectedSnsNeuronContext({
-      Component: SnsNeuronMetaInfoCard,
-      neuron,
-      reload: jest.fn(),
-    });
-
-    expect(queryByTestId("sns-increase-dissolve-delay")).toBeInTheDocument();
-  });
-
-  it("should not render dissolve button", async () => {
-    const neuron = mockSnsNeuronWithPermissions([]);
-    const { queryByTestId } = renderSelectedSnsNeuronContext({
-      Component: SnsNeuronMetaInfoCard,
-      neuron,
-      reload: jest.fn(),
-    });
-
-    expect(
-      queryByTestId("sns-increase-dissolve-delay")
-    ).not.toBeInTheDocument();
-  });
-
-  it("renders increase dissolve delay button", async () => {
-    const neuron = mockSnsNeuronWithPermissions([
-      SnsNeuronPermissionType.NEURON_PERMISSION_TYPE_CONFIGURE_DISSOLVE_STATE,
-    ]);
-    const { queryByTestId } = renderSelectedSnsNeuronContext({
-      Component: SnsNeuronMetaInfoCard,
-      neuron,
-      reload: jest.fn(),
-    });
-
-    expect(queryByTestId("sns-increase-dissolve-delay")).toBeInTheDocument();
-  });
-
-  it("should not render increase dissolve delay button", async () => {
-    const neuron = mockSnsNeuronWithPermissions([]);
-    const { queryByTestId } = renderSelectedSnsNeuronContext({
-      Component: SnsNeuronMetaInfoCard,
-      neuron,
-      reload: jest.fn(),
-    });
-
-    expect(
-      queryByTestId("sns-increase-dissolve-delay")
-    ).not.toBeInTheDocument();
-  });
+  // TODO: uncomment for display neuron age
+  // it("should render neuron age", () => {
+  //   const { getByTestId } = renderSelectedSnsNeuronContext({
+  //     Component: SnsNeuronMetaInfoCard,
+  //     neuron: mockSnsNeuron,
+  //     reload: jest.fn(),
+  //   });
+  //
+  //   expect(getByTestId("sns-neuron-age")?.textContent.trim()).toEqual(
+  //     secondsToDuration(BigInt(mockSnsNeuronTimestampSeconds))
+  //   );
+  // });
 });
