+++ frontend/src/tests/lib/utils/neuron.utils.spec.ts
@@ -1866,4 +1866,19 @@ describe("neuron-utils", () => {
       ).toBe(false);
     });
   });
+
+  // TODO: uncomment for display neuron age
+  // describe("neuronAge", () => {
+  //   it("should return ageSeconds property", () => {
+  //     expect(neuronAge(mockNeuron)).toEqual(mockNeuron.ageSeconds);
+  //   });
+  //
+  //   it("should return max age value", () => {
+  //     const neuron = {
+  //       ...mockNeuron,
+  //       ageSeconds: mockNeuron.ageSeconds + BigInt(SECONDS_IN_FOUR_YEARS),
+  //     };
+  //     expect(neuronAge(neuron)).toEqual(BigInt(SECONDS_IN_FOUR_YEARS));
+  //   });
+  // });
 });
